requirements.txt
pandas==2.1.4
numpy<2.0.0
python-dateutil>=2.8.2

scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.3

requests==2.31.0
hopsworks==3.4.4
urllib3<2.0.0

streamlit==1.30.0
plotly==5.18.0

python-dotenv==1.0.0